v2.5.0 Supporting also minimalistic html report
